switch to seperate folder and add todos for missing class descriptions
