+++ Integreat/Integreat.Shared/Models/Event/EventCategory.cs
@@ -1,0 +1,17 @@
+﻿using Newtonsoft.Json;
+
+namespace Integreat.Shared.Models
+{
+    /// <summary>
+    /// todo description
+    /// </summary>
+	public class EventCategory
+	{
+		[JsonProperty ("id")]
+		public int Id	{ get; set; }
+
+		[JsonProperty ("name")]
+		public string Name{ get; set; }
+	}
+}
+
