+++ Integreat/Integreat.Shared/Models/Event/EventTag.cs
@@ -1,0 +1,17 @@
+﻿using Newtonsoft.Json;
+
+namespace Integreat.Shared.Models
+{
+    /// <summary>
+    /// todo description
+    /// </summary>
+	public class EventTag
+	{
+		[JsonProperty ("name")]
+		public string Name { get; set; }
+
+		[JsonProperty ("id")]
+    	public int Id { get; set; }
+	}
+}
+
