--- Integreat/Integreat.Shared/Models/EventCategory.cs
@@ -1,16 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace Integreat.Shared.Models
-{
-	public class EventCategory
-	{
-
-		[JsonProperty ("id")]
-		public int Id	{ get; set; }
-
-		[JsonProperty ("name")]
-		public string Name{ get; set; }
-
-	}
-}
-
