--- Integreat/Integreat.Shared/Models/EventTag.cs
@@ -1,16 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace Integreat.Shared.Models
-{
-	public class EventTag
-	{
-
-		[JsonProperty ("name")]
-		public string Name { get; set; }
-
-		[JsonProperty ("id")]
-		public int Id { get; set; }
-
-	}
-}
-
